+++ lib/AppInfo/Application.php
@@ -1,7 +1,7 @@
 <?php
 /**
  *
- * (c) Copyright Ascensio System SIA 2025
+ * (c) Copyright Ascensio System SIA 2026
  *
  * This program is a free software product.
  * You can redistribute it and/or modify it under the terms of the GNU Affero General Public License
@@ -143,7 +143,7 @@ class Application extends App implements IBootstrap {
                     $templateManager->registerTemplateFileCreator(function () use ($appName, $trans) {
                         $wordTemplate = new TemplateFileCreator($appName, $trans->t("New document"), ".docx");
                         $wordTemplate->addMimetype("application/vnd.openxmlformats-officedocument.wordprocessingml.document");
-                        $wordTemplate->setIconClass("icon-onlyoffice-new-docx");
+                        $wordTemplate->setIconSvgInline(file_get_contents(__DIR__ . '/../../img/new-docx.svg'));
                         $wordTemplate->setRatio(21/29.7);
                         return $wordTemplate;
                     });
@@ -151,7 +151,7 @@ class Application extends App implements IBootstrap {
                     $templateManager->registerTemplateFileCreator(function () use ($appName, $trans) {
                         $cellTemplate = new TemplateFileCreator($appName, $trans->t("New spreadsheet"), ".xlsx");
                         $cellTemplate->addMimetype("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet");
-                        $cellTemplate->setIconClass("icon-onlyoffice-new-xlsx");
+                        $cellTemplate->setIconSvgInline(file_get_contents(__DIR__ . '/../../img/new-xlsx.svg'));
                         $cellTemplate->setRatio(21/29.7);
                         return $cellTemplate;
                     });
@@ -159,7 +159,7 @@ class Application extends App implements IBootstrap {
                     $templateManager->registerTemplateFileCreator(function () use ($appName, $trans) {
                         $slideTemplate = new TemplateFileCreator($appName, $trans->t("New presentation"), ".pptx");
                         $slideTemplate->addMimetype("application/vnd.openxmlformats-officedocument.presentationml.presentation");
-                        $slideTemplate->setIconClass("icon-onlyoffice-new-pptx");
+                        $slideTemplate->setIconSvgInline(file_get_contents(__DIR__ . '/../../img/new-pptx.svg'));
                         $slideTemplate->setRatio(16/9);
                         return $slideTemplate;
                     });