+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Change Log
+## Changed
+- replace deprecated TemplateFileCreator::setIconClass with TemplateFileCreator::setIconSvgInline method
 
 ##
 ## Changed