+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
-## Change
+
+##
+## Changed
+- fixed opening the editor with a watermark on tagged files
+- replace deprecated TemplateFileCreator::setIconClass with TemplateFileCreator::setIconSvgInline method
 - OCA\Onlyoffice\Preview class now implements OCP\Preview\IProviderV2 interface
 
 ## 9.12.0