+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ##
 ## Changed
+- fixed watermark not applying to link shares with system tags
 - fixed opening the editor with a watermark on tagged files
 - replace deprecated TemplateFileCreator::setIconClass with TemplateFileCreator::setIconSvgInline method
 - OCA\Onlyoffice\Preview class now implements OCP\Preview\IProviderV2 interface