+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-##
+## 10.0.0-beta
 ## Changed
 - compatible with Nextcloud 33
 - fixed watermark not applying to link shares with system tags